Move two IL edits into a single file with a static helper class.
